+++ handlers/batch-email-sender/src/main/scala/com/gu/batchemailsender/api/batchemail/BrazeSqsMessage.scala
@@ -1,0 +1,191 @@
+package com.gu.batchemailsender.api.batchemail
+
+import ai.x.play.json.Jsonx
+import SalesforceMessage.SalesforceBatchItem
+import SalesforceToBrazeTransformations._
+
+/**
+ * This is what actually gets sent to SQS and the fields correspond to Braze api_trigger_properties.
+ * For example, stopped_credit_summaries can be found in SV_HolidayStopConfirmation_GuardianWeeklyBodyCopy
+ * Braze content block.
+ *
+    {
+      "To" : {
+        "Address" : "mario.galic+bvtgedltoa@guardian.co.uk",
+        "SubscriberKey" : "mario.galic+bvtgedltoa@guardian.co.uk",
+        "ContactAttributes" : {
+          "SubscriberAttributes" : {
+            "first_name" : "bvtgedltoa",
+            "last_name" : "bvtgedltoa",
+            "subscriber_id" : "",
+            "product" : "",
+            "delivery_address_change_line1" : "Address line 1",
+            "delivery_address_change_line2" : "Address line 2",
+            "delivery_address_change_city" : "city",
+            "delivery_address_change_state" : "state",
+            "delivery_address_change_postcode" : "postcode",
+            "delivery_address_change_country" : "Antigua & Barbuda",
+            "delivery_address_change_effective_date_blurb" : "Guardian weekly subscription (A-S00060454)  as of front cover dated Friday 10th April 2020\n\n(as displayed on confirmation page at 12:39:44  on 26th March 2020)"
+          }
+        }
+      },
+      "DataExtensionName" : "SV_DeliveryAddressChangeConfirmation",
+      "SfContactId" : "0033E00001Chmk9QAB",
+      "IdentityUserId" : "200002073"
+    }
+ *
+ * FIXME: For some reason this has to have flat structure?
+ */
+case class BrazeApiTriggerProperties(
+  first_name: String,
+  last_name: String,
+  subscriber_id: String,
+  next_charge_date: Option[String],
+  product: String,
+  modified_by_customer: Option[Boolean],
+  holiday_start_date: Option[String],
+  holiday_end_date: Option[String],
+  stopped_credit_sum: Option[String],
+  currency_symbol: Option[String],
+  stopped_issue_count: Option[String],
+  stopped_credit_summaries: Option[List[StoppedCreditSummary]],
+  digital_voucher: Option[DigitalVoucher],
+
+  /*
+   * Delivery Problem fields
+   * Braze campaign: SV_DeliveryProblemConfirmation
+   * Braze content block: SV_HolidayStopConfirmation_GuardianWeeklyBodyCopy
+   */
+  delivery_problem_action: Option[String] = None, // Credit or Escalation
+  delivery_problem_total_affected: Option[String] = None,
+  delivery_problem_affected_dates: Option[String] = None,
+  delivery_problem_total_credit: Option[String] = None,
+  delivery_problem_invoice_date: Option[String] = None, // Invoice_Date__c Is it the same for multiple issues ?
+  delivery_problem_type: Option[String] = None, // No delivery | Damaged paper,
+  delivery_problem_currency_symbol: Option[String] = None,
+  delivery_problem_case_number: Option[String] = None,
+
+  /*
+   * SV_DeliveryAddressChangeConfirmation
+   * https://manage.theguardian.com/delivery/guardianweekly/address/confirmed
+   */
+  delivery_address_change_line1: Option[String] = None,
+  delivery_address_change_line2: Option[String] = None,
+  delivery_address_change_city: Option[String] = None,
+  delivery_address_change_state: Option[String] = None,
+  delivery_address_change_postcode: Option[String] = None,
+  delivery_address_change_country: Option[String] = None,
+  delivery_address_change_effective_date_blurb: Option[String] = None,
+)
+
+case class StoppedCreditSummary(credit_amount: Double, credit_date: String)
+case class DigitalVoucherUrl(value: String) extends AnyVal
+case class DigitalVoucher(barcodeUrl: DigitalVoucherUrl)
+case class EmailPayloadContactAttributes(SubscriberAttributes: BrazeApiTriggerProperties)
+case class EmailPayloadTo(Address: String, SubscriberKey: String, ContactAttributes: EmailPayloadContactAttributes)
+
+// Message put on the SQS contributions-thanks queue for pickup by membership-workflow
+case class BrazeSqsMessage(To: EmailPayloadTo, DataExtensionName: String, SfContactId: Option[String], IdentityUserId: Option[String], recordId: String)
+
+object BrazeSqsMessage {
+  implicit val bb = Jsonx.formatCaseClass[StoppedCreditSummary]
+  implicit val cc = Jsonx.formatCaseClass[DigitalVoucherUrl]
+  implicit val uu = Jsonx.formatCaseClass[DigitalVoucher]
+  implicit val emailPayloadSubscriberAttributesWriter = Jsonx.formatCaseClass[BrazeApiTriggerProperties]
+  implicit val emailPayloadContactAttributesWriter = Jsonx.formatCaseClass[EmailPayloadContactAttributes]
+  implicit val emailPayloadToWriter = Jsonx.formatCaseClass[EmailPayloadTo]
+  implicit val emailToSendWriter = Jsonx.formatCaseClass[BrazeSqsMessage]
+
+  def fromSalesforceMessage(salesforceBatchItem: SalesforceBatchItem): BrazeSqsMessage = {
+    val salesforcePayload = salesforceBatchItem.payload
+
+    val emailPayloadTo = EmailPayloadTo(
+      Address = salesforceBatchItem.payload.to_address,
+      SubscriberKey = salesforceBatchItem.payload.to_address,
+
+      // Actual Braze fields
+      ContactAttributes = EmailPayloadContactAttributes(
+        SubscriberAttributes = BrazeApiTriggerProperties(
+          salesforceBatchItem.payload.first_name,
+          salesforceBatchItem.payload.last_name,
+          salesforceBatchItem.payload.subscriber_id,
+          salesforceBatchItem.payload.next_charge_date,
+          salesforceBatchItem.payload.product,
+          salesforceBatchItem.payload.modified_by_customer,
+
+          // Holiday stop
+          salesforcePayload.holiday_stop_request.map(_.holiday_start_date),
+          salesforcePayload.holiday_stop_request.map(_.holiday_end_date),
+          salesforcePayload.holiday_stop_request.map(_.stopped_credit_sum),
+          salesforcePayload.holiday_stop_request.map(_.currency_symbol),
+          salesforcePayload.holiday_stop_request.map(_.stopped_issue_count),
+          stopped_credit_summaries = for {
+            stop <- salesforcePayload.holiday_stop_request
+            summaryList <- stop.stopped_credit_summaries
+            stoppedCreditSummaries = summaryList.map(detail => StoppedCreditSummary(detail.credit_amount, fromSfDateToDisplayDate(detail.credit_date)))
+          } yield stoppedCreditSummaries,
+
+          // Digital voucher
+          digital_voucher = salesforceBatchItem
+            .payload
+            .digital_voucher
+            .map(wireVoucher => DigitalVoucher(DigitalVoucherUrl(wireVoucher.barcode_url))),
+
+          // Delivery Problem
+          delivery_problem_action = salesforcePayload.delivery_problem.map(_.actionTaken),
+          delivery_problem_total_affected = salesforcePayload.delivery_problem.map(_.issuesAffected),
+          delivery_problem_affected_dates = salesforcePayload.delivery_problem.map(_.deliveries.map(d => fromSfDateToDisplayDate(d.Delivery_Date__c)).mkString(", ")),
+          delivery_problem_total_credit = salesforcePayload.delivery_problem.map(_.totalCreditAmount),
+          delivery_problem_invoice_date = salesforcePayload.delivery_problem.flatMap(_.deliveries.flatMap(_.Invoice_Date__c).headOption.map(fromSfDateToDisplayDate)),
+          delivery_problem_type = salesforcePayload.delivery_problem.map(_.typeOfProblem),
+          delivery_problem_currency_symbol = salesforcePayload.delivery_problem.map(_.currencySymbol),
+          delivery_problem_case_number = salesforcePayload.delivery_problem.map(_.caseNumber),
+
+          // Delivery Address Change
+          delivery_address_change_line1 = salesforcePayload.delivery_address_change.flatMap(_.mailingStreet).flatMap(sfStreetToLine1),
+          delivery_address_change_line2 = salesforcePayload.delivery_address_change.flatMap(_.mailingStreet).flatMap(sfStreetToLine2),
+          delivery_address_change_city = salesforcePayload.delivery_address_change.flatMap(_.mailingCity),
+          delivery_address_change_state = salesforcePayload.delivery_address_change.flatMap(_.mailingState),
+          delivery_address_change_postcode = salesforcePayload.delivery_address_change.flatMap(_.mailingPostalCode),
+          delivery_address_change_country = salesforcePayload.delivery_address_change.flatMap(_.mailingCountry),
+          delivery_address_change_effective_date_blurb = salesforcePayload.delivery_address_change.flatMap(_.addressChangeEffectiveDateBlurb),
+        )
+      )
+    )
+
+    BrazeSqsMessage(
+      To = emailPayloadTo,
+      DataExtensionName = brazeCampaignId(salesforceBatchItem),
+      SfContactId = Some(salesforcePayload.sf_contact_id),
+      IdentityUserId = salesforcePayload.identity_id,
+      recordId = salesforcePayload.record_id
+    )
+  }
+
+  // This is mapped to Braze Template API Identifier by membership-workflow
+  // https://github.com/guardian/membership-workflow/blob/2e354b81888f6d222d9de0b4c2eda8e0f2b14729/app/services/BrazeTemplateLookupService.scala#L13
+  def brazeCampaignId(salesforceBatchItem: SalesforceBatchItem): String =
+    (salesforceBatchItem.object_name, salesforceBatchItem.payload.email_stage) match {
+      case ("Card_Expiry__c", _) => "expired-card"
+      case ("DD_Mandate_Failure__c", "MF1") => "dd-mandate-failure-1"
+      case ("DD_Mandate_Failure__c", "MF2") => "dd-mandate-failure-2"
+      case ("DD_Mandate_Failure__c", "MF3") => "dd-mandate-failure-3"
+      case ("DD_Mandate_Failure__c", "MF4") => "dd-mandate-failure-4"
+      case ("DD_Mandate_Failure__c", "MF5") => "dd-mandate-failure-5"
+      case ("DD_Mandate_Failure__c", "MF6") => "dd-mandate-failure-6"
+      case ("DD_Mandate_Failure__c", "MF7") => "dd-mandate-failure-7"
+      case ("DD_Mandate_Failure__c", "MF8") => "dd-mandate-failure-8"
+      case ("Payment_Failure__c", "DD_PF1") => "SV_DDpaymentfailure1"
+      case ("Payment_Failure__c", "DD_PF2") => "SV_DDpaymentfailure2"
+      case ("Payment_Failure__c", "DD_PF3") => "SV_DDpaymentfailure3"
+      case ("Payment_Failure__c", "DD_PF4") => "SV_DDpaymentfailure4"
+      case ("Holiday_Stop_Request__c", "create") => "SV_HolidayStopConfirmation"
+      case ("Holiday_Stop_Request__c", "amend") => "SV_HolidayStopAmend"
+      case ("Holiday_Stop_Request__c", "withdraw") => "SV_HolidayStopWithdrawal"
+      case ("Digital_Voucher__c", "create") => "SV_VO_NewCard"
+      case ("Digital_Voucher__c", "replace") => "SV_VO_ReplacementCard"
+      case ("Case", "Delivery issues") => "SV_DeliveryProblemConfirmation"
+      case ("Contact", "Delivery address change") => "SV_DeliveryAddressChangeConfirmation"
+      case (objectName, emailStage) => throw new RuntimeException(s"Unrecognized (object_name, email_stage) = ($objectName, $emailStage). Please fix SF trigger.")
+    }
+}
